+++ src/entidades/Avion.py
@@ -1,8 +1,9 @@
 from typing import Any
+from decimal import Decimal
 
 class AvionBase:
 
-    def __init__(self, matricula: str, marca: str, modelo: str, capacidad: str, autonomia_km: str, costo_hora_vuelo: float, id_estado_actual: int) -> None:
+    def __init__(self, matricula: str, marca: str, modelo: str, capacidad: str, autonomia_km: str, costo_hora_vuelo: Decimal, id_estado_actual: int) -> None:
         self.matricula = matricula
         self.marca = marca
         self.modelo = modelo
@@ -27,7 +28,7 @@ class AvionBase:
 
 class AvionDesdeDB(AvionBase):
 
-    def __init__(self, id: int, matricula: str, marca: str, modelo: str, capacidad: str, autonomia_km: str, costo_hora_vuelo: float, id_estado_actual: int) -> None:
+    def __init__(self, id: int, matricula: str, marca: str, modelo: str, capacidad: str, autonomia_km: str, costo_hora_vuelo: Decimal, id_estado_actual: int) -> None:
         super().__init__(matricula, marca, modelo, capacidad, autonomia_km, costo_hora_vuelo, id_estado_actual)
         self.id = id
